requirements.txt
Pillow==9.5.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
tkcalendar==1.6.1
mysql-connector-python==8.0.33 